requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.7.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
blinker=1.8.2=pypi_0
blis=0.7.11=pypi_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2024.7.4=h56e8100_0
catalogue=2.0.10=pypi_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.18.1=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
confection=0.1.5=pypi_0
cymem=2.0.8=pypi_0
debugpy=1.8.2=py312h275cf98_0
decorator=5.1.1=pyhd8ed1ab_0
en-core-web-sm=3.7.1=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
ffmpeg-python=0.2.0=pypi_0
filelock=3.15.4=pypi_0
flask=3.0.3=pypi_0
fsspec=2024.6.1=pypi_0
future=1.0.0=pypi_0
huggingface-hub=0.23.5=pypi_0
idna=3.7=pypi_0
importlib-metadata=8.0.0=pyha770c72_0
importlib_metadata=8.0.0=hd8ed1ab_0
intel-openmp=2021.4.0=pypi_0
ipykernel=6.29.5=pyh4bbf305_0
ipython=8.26.0=pyh7428d3b_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h2e8e312_0
krb5=1.21.3=hdf4eb48_0
langcodes=3.4.0=pypi_0
language-data=1.2.0=pypi_0
libexpat=2.6.2=h63175ca_0
libffi=3.4.2=h8ffe710_5
libsodium=1.0.18=h8d14728_1
libsqlite=3.46.0=h2466b09_0
libzlib=1.3.1=h2466b09_1
marisa-trie=1.2.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=pypi_0
mpmath=1.3.0=pypi_0
murmurhash=1.0.10=pypi_0
mysql-connector-python=9.0.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
nltk=3.8.1=pypi_0
numpy=1.26.4=pypi_0
openssl=3.3.1=h2466b09_2
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.2=pyhd8ed1ab_0
preshed=3.0.9=pypi_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=6.0.0=py312h4389bb4_0
pure_eval=0.2.2=pyhd8ed1ab_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
python=3.12.4=h889d299_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pytz=2024.1=pypi_0
pywin32=306=py312h53d5487_2
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=py312hd7027bb_0
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.14.0=pypi_0
setuptools=70.3.0=pyhd8ed1ab_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyh6c4a22f_0
smart-open=7.0.4=pypi_0
spacy=3.7.5=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
spacy-lookups-data=1.0.5=pypi_0
speechrecognition=3.10.4=pypi_0
srsly=2.4.8=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tbb=2021.11.0=pypi_0
thinc=8.2.5=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=h5226925_1
tokenizers=0.19.1=pypi_0
torch=2.3.1+cu121=pypi_0
torchaudio=2.3.1+cu121=pypi_0
torchvision=0.18.1+cu121=pypi_0
tornado=6.4.1=py312h4389bb4_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.42.4=pypi_0
typer=0.12.3=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.2.2=pypi_0
vc=14.3=h8a93ad2_20
vc14_runtime=14.40.33810=ha82c5b3_20
vs2015_runtime=14.40.33810=h3bf8584_20
wasabi=1.1.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
weasel=0.4.1=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
wrapt=1.16.0=pypi_0
xz=5.2.6=h8d14728_0
zeromq=4.3.5=he1f189c_4
zipp=3.19.2=pyhd8ed1ab_0
